improved code readability and responsive fixes
